feat:  Multiple Datasources Test for the RDBMS module using test templates #21983
- Moved RDBMS ITests to global QA Folder
- Fix Liquibase script for timestamp in different database

+++ db/rdbms/src/main/java/io/camunda/db/rdbms/read/service/ProcessInstanceReader.java
@@ -34,7 +34,7 @@ public class ProcessInstanceReader {
     LOG.trace("[RDBMS DB] Search for process instance with filter {}", filter);
     final var totalHits = processInstanceMapper.count(filter);
     final var hits = processInstanceMapper.search(filter);
-    return new SearchResult(hits, totalHits);
+    return new SearchResult(hits, totalHits.intValue());
   }
 
   public record SearchResult(List<ProcessInstanceEntity> hits, Integer total) {}
